Multi-copter control, camera calibration, image<->world coordinate transform, camera identification using vendor and product IDs, and more!
Huuge update! Here's a summary of the main changes:
 - Re-structured the main file (full_control_with_cam) to allow for any
number of drones. I've split the task into 2 classes: Spotter (vision
code) and WorkerDrone (radio tx/rx, logging, etc. for each individual
CF).
 - Added code to calibrate camera(s), so that we can account for lens
distortion effects (improved localization especially near the image
edges, where distortion is greater).
 - Also incorporated image to/from world coordinate transform,
including the possibility to estimate camera's pose
(position+orientation) wrt the world using a calibration pattern. This
allows us to correctly convert pixel coordinates to world coordinates,
and greatly improves the accuracy when the camera is not perfectly
horizontal (before, we would just assume it was horizontal so that
vertical axis in image space corresponded to up-down z-axis in the real
world, but that was only true for the horizontal case). These functions
are in vision_aux_functions, and are already included in the Spotter
code to accurately control the drones.
 - All new uvc cameras I bought from amazon have the same camera name,
so I had to add vendor & product ID identification so I could tell them
apart.
 - Deleted some old files (test_*).