Better velocity estimation and minor fixes
Main contribution of this commit is the addition of a derivative filter
which estimates much more accurately velocity and acceleration from the
worker's position in the spotter's camera. It's located in
plot_tools/DerivativeHelper.
Other minor fixes are:
 - Incorporated derivative filter into full_control_with_cam.py.
 - Fixed Markdown syntax in the Readme.md so code fragments show up
correctly. Also added SDL2 as a requirement.
 - uvc_capture would incorrectly show camera name information, now it's
fixed
 - Fixed some symlinks
 - Added a modified version of full_control_with_cam with code that
logs and computes acceleration of the worker measured from its IMU as
well as derivated from its position seen from the spotter's camera, for
IPSN poster (which just got accepted)